Move RealCommandRunner into its own file
