+++ configure.py
@@ -549,6 +549,7 @@ for name in ['build',
              'metrics',
              'missing_deps',
              'parser',
+             'real_command_runner',
              'state',
              'status_printer',
              'string_piece_util',
